refactor(backend): Splits seeding
